[refactor] : Separate the "validation" functionality into a "locationValidator" class.
